requirements.txt
fastapi==0.105.0
PyJWT==2.8.0
PyMySQL==1.1.0
uvicorn==0.25.0
pydantic~=2.5.3
bcrypt~=4.1.2
resend==0.6.0
firebase-admin==6.4.0
openai==1.55.3
Pyrebase4==4.8.0
setuptools==75.8.0
